Elimintate dead cycles in GCD input/output
Add skid buffers on the input and output of GCD, so we can further
decouple the state machine from the input and output.

This version runs almost twice as fast as the original when there's
no delays on the input and no backpressure on the output.  It runs
more than twice as fast as the original when there's delay on the
input or backpressure on the output, as it overlaps the compute with
those delays.

Original:

[115] cycles: 650  Delay: in=0 out=0
[115] cycles: 750  Delay: in=0 out=1
[115] cycles: 850  Delay: in=0 out=2
[115] cycles: 749  Delay: in=1 out=0
[115] cycles: 849  Delay: in=1 out=1
[115] cycles: 949  Delay: in=1 out=2
[115] cycles: 848  Delay: in=2 out=0
[115] cycles: 948  Delay: in=2 out=1
[115] cycles: 1048  Delay: in=2 out=2

This version:

[115] cycles: 371  Delay: in=0 out=0
[115] cycles: 399  Delay: in=0 out=1
[115] cycles: 436  Delay: in=0 out=2
[115] cycles: 385  Delay: in=1 out=0
[115] cycles: 400  Delay: in=1 out=1
[115] cycles: 436  Delay: in=1 out=2
[115] cycles: 417  Delay: in=2 out=0
[115] cycles: 419  Delay: in=2 out=1
[115] cycles: 438  Delay: in=2 out=2